requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
faiss-cpu
unstructured
html2text
beautifulsoup4
lxml
python-dotenv
fastembed
google-generativeai
